Fixed bug with invalid id assignation on alarm creation. Alarm id type migrated to long

+++ app/src/main/java/sleepless_nights/location_alarm/alarm/view_models/AlarmViewModel.java
@@ -38,7 +38,7 @@ public class AlarmViewModel extends AndroidViewModel {
     }
 
     @Nullable
-    public Alarm getAlarmLiveDataById(int id) {
+    public Alarm getAlarmLiveDataById(long id) {
         return alarmRepository.getAlarmById(id);
     }
